Hash password
Added a system to create a hash of the password before storing it on the DataBase.

Now, when creating an account, you don't have to insert the passwortd, because of the first password is auto-generated.

You can't connect with the initial password, but the API will notice you to change it in the new route.

1 route added for "mudarsenha" change password.

Token duration time changed from 3 hours to 1 hour